COM-2606 - moves getAbsenceHours from historyExport helper to draftPay helper
